requirements.txt
Flask==2.3.2
firebase-admin google-cloud-storage
pyrebase4 --target=./